requirements.txt
fastapi
uvicorn
openai
tiktoken
faiss-cpu
python-dotenv
pypdf
numpy
sentence-transformers